Fix bug in `remove_all` where size was not decreased
In the `remove_all` function, the size of the multiset was not altered. It
should have been decreased by the number of elements removed. This commit adds
this behavior to properly maintain the size of the multiset.

In addition, this commit adds tests in the `remove_all` documentation comment
test and a more extended test in the test_multiset module.