Implement union, intersection, union_counts, and intersection_counts
The `union` and `union_counts` methods return iterators that yield elements the
maximum number of times they appear in either set. This is different from the
previously implemented `Add` trait, which yields each element the combined
number of times it appears in both sets.

The `intersection` and `intersection_counts` methods return iterators that yield
elements the minimum number of times they appear in either set. This is distinct
from the previously implemented `Sub` trait, which yields each element the
number of times it appears in the first set less the number of times it appears
in the second.

The different between the plain variants and the `_counts` variants is that the
former yields only the elements (repeated the correct number of times) while the
latter yields a tuple of an element and the count of that element.